requirements.txt
pygame==1.9.3
av==0.4.0
Adafruit_GPIO==1.0.3
Pillow==5.3.0
